Move files, update types and refactor
