+++ projects/ngx-testing-tools/src/lib/common/expectation/should-create.ts
@@ -1,6 +1,5 @@
 export function shouldCreate(factory: () => unknown): void {
   it('should create', () => {
-    const instance: unknown = factory();
-    expect(instance).toBeTruthy();
+    expect(factory()).toBeTruthy();
   });
 }
