--- projects/ngx-testing-tools/src/lib/common/test-bed/models/common-setup.model.ts
@@ -1,3 +1,0 @@
-import { CommonTools } from './common-tools.model';
-
-export type CommonSetup<I extends {}> = (tools: CommonTools<I>, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
