--- projects/ngx-testing-tools/src/lib/common/test-bed/models/common-test-bed.model.ts
@@ -1,1 +1,0 @@
-export interface CommonTestBed<T, R> {}
