+++ projects/ngx-testing-tools/src/lib/common/test-bed/models/common-tools.model.ts
@@ -3,7 +3,7 @@ import { InjectionStore } from '../../../components';
 
 export interface CommonTools<I extends {} = {}> extends InjectionStore<I> {
   /**
-   * The fixture injector.
+   * The root injector.
    */
   injector: Injector;
 }
