+++ projects/ngx-testing-tools/src/lib/common/test-bed/models/custom-test-bed.model.ts
@@ -1,0 +1,8 @@
+import { PrettyMerge } from '../../../models/shared.model';
+import { CommonTestBedFactory } from '../common-test-bed-factory';
+
+export type CustomTestBed<
+  Fn extends (...args: any[]) => jasmine.ImplementationCallback,
+  Factory extends CommonTestBedFactory<any>
+> = PrettyMerge<Fn & Factory>
+
