+++ projects/ngx-testing-tools/src/lib/common/test-bed/models/enhanced-callback.model.ts
@@ -1,0 +1,3 @@
+import { CommonTools } from './common-tools.model';
+
+export type EnhancedCallback<Tools extends CommonTools> = (tools: Tools, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
