+++ projects/ngx-testing-tools/src/lib/common/test-bed/store/models/injected-store.model.ts
