+++ projects/ngx-testing-tools/src/lib/components/test-bed/assertions/assert-component.ts
@@ -1,6 +1,6 @@
 import { Component, Type } from '@angular/core';
-import { Nullable } from '../../models/shared.model';
-import { isComponentAnnotation } from './component-annotation';
+import { Nullable } from '../../../models/shared.model';
+import { isComponentAnnotation } from '../component-annotation';
 
 export function assertComponent(TypeCtor: Type<any>, annotation: Nullable<Component>): void {
   if (!isComponentAnnotation(annotation))
