+++ projects/ngx-testing-tools/src/lib/components/test-bed/component-tools.ts
@@ -1,14 +1,14 @@
 import { ProviderToken } from '@angular/core';
 import { ComponentFixture } from '@angular/core/testing';
-import { assertComponentFixture } from './assert-fixture';
+import { InjectionStore } from '../../common/test-bed/store';
+import { buildInjected } from '../../common/test-bed/store/injected';
+import { assertComponentFixture } from './assertions/assert-fixture';
 import { buildComponentActionTools } from './component-action-tools';
 import { buildComponentQueryTools } from './component-query-tools';
 import { ComponentTestBedFactory } from './component-test-bed-factory';
 import { ComponentTools } from './models';
 import { ComponentActionTools } from './models/component-action-tools.model';
 import { ComponentQueryTools } from './models/component-query-tools.model';
-import { InjectionStore } from './store';
-import { buildInjected } from './store/injected';
 
 export function buildComponentTools(factory: ComponentTestBedFactory<any>): ComponentTools<any, any> {
   const fixture: ComponentFixture<unknown> = factory['fixture'];
