+++ projects/ngx-testing-tools/src/lib/components/test-bed/index.ts
@@ -1,3 +1,3 @@
 export * from './models';
-export * from './store';
+export * from '../../common/test-bed/store';
 export { componentTestBed } from './component-test-bed';
