+++ projects/ngx-testing-tools/src/lib/components/test-bed/models/component-setup.model.ts
@@ -1,0 +1,4 @@
+import { EnhancedCallback } from '../../../common/test-bed/models/enhanced-callback.model';
+import { ComponentTools } from './component-tools.model';
+
+export type ComponentSetup<T, I extends {}> = EnhancedCallback<ComponentTools<T, I>>
