--- projects/ngx-testing-tools/src/lib/components/test-bed/models/setup-fn.model.ts
@@ -1,3 +1,0 @@
-import { ComponentTools } from './component-tools.model';
-
-export type ComponentSetup<T, I extends {}> = (tools: ComponentTools<T, I>, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
