+++ projects/ngx-testing-tools/src/lib/service/test-bed/models/service-setup.model.ts
@@ -1,3 +1,3 @@
 import { ServiceTools } from './service-tools.model';
 
-export type ServiceSetup<T, I> = (tools: ServiceTools<T, I>, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
+export type ServiceSetup<T, I extends {}> = (tools: ServiceTools<T, I>, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
