+++ projects/ngx-testing-tools/src/lib/service/test-bed/models/service-tools.model.ts
@@ -1,5 +1,8 @@
-import { InjectionStore } from '../../../components';
+import { CommonTools } from '../../../common/test-bed/models/common-tools.model';
 
-export interface ServiceTools<T, I> extends InjectionStore<I> {
+export interface ServiceTools<T, I extends {}> extends CommonTools<I> {
+  /**
+   * Service instance.
+   */
   service: T;
 }
